Added collapsed wasted effort calculations.
[Details]
Added the option to calculate wasted effort based on the groups
(spectrally identical elements), instead of on the elements themselves.
This is useful in the case of Daikon, where with a bad test suite, we
have large groups of spectrally identical elements. In this case, moving
these groups around with the feedback localizer causes large changes in
the wasted effort, making the feedback localizer appear to be doing
worse. However, if wasted effort is only considered for each group, it
can be seen that drastically less groups are needed to be inspected
using the feedback localizer.

[File's changed]
	modified:   feedback_loc.py
	modified:   weffort.py