Added feedback outer loop formalization
[Details]
Before we had the feedback "multi", which represented doing the feedback
loop multiple times with certain elements not considered. This was
however, just a heuristic, and did not change the result much. This was
significantly changed in this commit to be the multi-fault/-explanation
iteration, which allows the proper processing of dominators, and also
allows alternative explanations to be given by the feedback localizer.

[Note]
The code for giving the previous faulty elements to the feedback
localizer still needs to be removed, although this will not change the
functionality of the code.

[Bugs]
A small bug was fixed in the suspiciousness calculation for tarantula,
where if given no passing test cases it would throw an exception.

[File's changed]

	modified:   feedback_loc.py
	modified:   suspicious.py