Added the otie tie break method.
[Details]
Added another tie break method called original ranking tie break (otie).
This method uses the original localizers scores to break ties further
down in the feedback localizer. Initial experiments show that this does
not help that much in improving the feedback localizers wasted effort,
but it does not worsen the effort by much either. This is a good result
because it shows that the variance in the ties is not large.

[File's changed]

	modified:   feedback_loc.py
	modified:   localize.py
	modified:   merge_rndm
	modified:   run_all