Added various running updates and multi2.
[Details]
Added the option to not have multi enabled at all (in the form of multi
and multi2). Also changed a number of things with the running scripts
used for the experiments. Of these, one thing that was changed was to
have the default tie mechanism being break by original ranking, then
original execution counts.

[Bugs]
There was a bug in the tie breaking where the ties were broken first by
execution count and then by original ranking.

[File's changed]

	modified:   feedback_loc.py
	modified:   localize.py
	modified:   merge.py
	modified:   plot.py
	modified:   run_all