Changed fault identification to multi-location.
[Details]
The major change in this commit was to add support for multi-location
faults in the identification and evaluation of the fault finding. This
included new input format that is backwards compatible, but now supports
multi-location faults with the use of integer numbering for each fault.
Identified faults are now stored in a python dict which is indexed by
the general fault number and gives an array of locations. Modified the
wasted effort calculations to include support for multi-location faults
as well.
Also completely reworked the find_unex method to not only find the
unexposed faults, but to also find the equivalence classes for a single
multi-location fault, so that these can be split.

[Note]
TODO: still need to adapt the top calculations for this.
TODO: still need to finish finding the split for bugs

[File's changed]

	modified:   feedback_loc.py
	modified:   find_unex.py
	modified:   input.py
	modified:   tcm_input.py
	modified:   top.py
	modified:   weffort.py