Fixed bugs with perc@n and added recursive merge.
[Details]
Added a recursive merge function that allows the merge script to be
called in the root directory of a multi-bug project version, and have
the results accumulated over all the fault numbers. This is useful for
the percent_at_n script, where results accumulated over all number of
faults is desired.

[Bugs]
Fixed two bugs in the percent_at_n.py script:
  * Fixed an off by one error in the calculation that was an artifact of
  the copied code from wasted effort calculations
  * Fixed an issue with duplicate min values in the combining function.
  Duplicated values are now merged into one.

[File's changed]

	modified:   merge.py
	modified:   percent_at_n.py
	modified:   run_all