Fixed bugs, added rec merge and auc calculation.
[Details]
Added a true recursive merge function that allows you to either specify
a recursion depth or recurse until "-fault" is found.
Also finished up the Area Under Curve calculations.
Fixed a bug in the wasted effort calculation that needs rerunning.

[Bugs]
There is still a bug in the precision and recall scripts that was due to
an oversight hack that is not yet fixed. This is to do with calculating
the expected values in groups.

[File's changed]

	modified:   merge.py
	modified:   percent_at_n.py
	modified:   plot.py
	modified:   weffort.py