Added plot all for perc@n, updated prec/rec calc.
[Details]
Added the abillity to plot a percentage at n graph with all the metrics
in one plot. Also updated the precision/recall calculation to be the
accurate combinatoric calculation that hopefully doesn't explode...
Finally, minor changes to what is run in the all mode.

[Bugs]
The precision/recall bug was fixed in this revision.

[File's changed]

	modified:   feedback_loc.py
	modified:   merge.py
	modified:   percent_at_n.py
	modified:   precision_recall.py