Refactored metric calculations and added new ones
[Details]
Added all of the metrics given in the paper: "An evaluation of pure
spectrum-based fault localization techniques for large-scale software
systems". Also refactored the way the calculations are called to be more
generic. Added a new metric is now as simple as adding a new function to
the suspicious.py file. Fixed up the merge and run_all scripts to deal
with this.

[File's changed]

	modified:   feedback_loc.py
	new file:   get_metric_names.py
	modified:   localize.py
	modified:   merge.py
	modified:   run_all
	modified:   suspicious.py