Renamed everything to flitsr, made more usable.
[Details]
Renamed the "feedback"'s to flitsr's, and refactored everything that
used that name. Also made the framework more usable by allowing a
environment variable to be specified that will work for the scripts.

[Note]
Even old and unused scripts were edited to use the new format. This was
done for compatibility in case these scripts are revived at any point.

[File's changed]

	modified:   average
	modified:   delete
	deleted:    feedback
	new file:   flitsr
	renamed:    feedback_loc.py -> flitsr.py
	modified:   merge
	modified:   merge.py
	modified:   move
	modified:   parallel.py
	modified:   percent_at_n
	modified:   percent_at_n.py
	modified:   plot
	modified:   plot.py
	modified:   reduce
	modified:   reduce_all
	modified:   run_all