Major refactor of input/output formats.
[Details]
Made the outputs of the rankings for both TCM and Gzoltar input methods
look identical (as much as possible). Also removed the Gzoltar input
method and made that input type use the same details object as the TCM
input type. Because of this change, it was possible to refactor the
outputting of either input type to a new output.py file, where the
functions are shared between the different input methods.

[File's changed]

	modified:   flitsr.py
	modified:   input.py
	new file:   output.py
	modified:   tcm_input.py