Added method level evaluation.
[Details]
Added the necessary code in the input files that allows the flitsr
framework to read in a statement level evaluation and convert it to a
method level evaluation on the fly. This is one by summing up the
contributions for each of the statements in the method, available in
the spectra. Once these contributions are collected, the individual
statements are discarded and only the methods remain (which will then be
grouped accordingly).

[Bugs]
There may be some minor bugs, especially in the input.py file which was
implemented quickly and based off of the TCM approach.

[File's changed]

	modified:   flitsr.py
	modified:   input.py
	modified:   tcm_input.py