Adapted cutoffs to account for multi-line faults.
[Details]
The cutoffs now respect multi-line faults in that if a line from a
multi-line fault has already been seen, the rest of the lines are no
longer considered for changing the cutoff point.

[Note]
This does not handle the worst case for multi-line bugs, where the
developer must look at all lines before understanding the bug.

[File's changed]

	modified:   cutoff_points.py
	modified:   flitsr.py