Fixed up parallel script and refactored.
[Details]
Fixed up the parallel implementation by recompiling a new version of the
jar file that supports the msp and bdm algorithms, and then editing the
flitsr scripts that call this. The changes to the flitsr scripts
included:
  - allowing for the different algorithms
  - removing elements from the partitioned spectra that are not in the
    failing test cases in that block
  - Allowing all flitsr options to be applied to the parallel script

Also refactored the code which helps with these parallel changes. The
refactoring moved cutoff point calculations to their own function, split
the run and output scripts, and allowed the execution of all these
scripts multiple times which is needed for the parallel computations.

[Bugs]
Fixed a semi-major bug in the sorting where the original ranking was
being preserved between executions.

[File's changed]

	modified:   flitsr.py
	modified:   parallel-1.0-SNAPSHOT-jar-with-dependencies.jar
	modified:   parallel.py