Added ranking reading and fixed basis cutoff.
[Details]
Added option to read in a ranking in GZoltar format instead of a
spectra, which can then be used to evaluate the ranking in the same way.
Useful for metrics that can be used in GZoltar to produce rankings (such
as barinel).

[Bugs]
Also fixed a bug in the basis cutoff calcuation and set worst effort in
cutoff to default to false.

[File's changed]

	modified:   cutoff_points.py
	modified:   flitsr.py
	new file:   ranking.py