Removed tcm option and fixed flitsr crash.
[Details]
Removed the required tcm option from FLITSR's execution, and instead
made FLITSR automatically detect which coverage type is being given.
Kept the tcm option as valid (but not doing anything) for backwards
compatibility.
Also stopped FLITSR from crashing when there is a test case that does
not execute any statements. Made FLITSR print out a warning in this case
instead, and continue to construct the basis.

[File's changed]

	modified:   ../README.md
	modified:   ../run_all
	modified:   flitsr.py