Changed the ranking reader to read flitsr output.
[Details]
Changed the default ranking reader to read in the output produced by the
flitsr program. This allows any saved rank to be read in by flitsr
again, and other metrics computed over this rank.

[Note]
The default reading in of a Gzoltar ranking that was there before this
version was replaced. This reader is still available, and may be
re-enabled in a future version.

[File's changed]

	modified:   flitsr.py
	modified:   ranking.py