Feature: FLITSR can now read any ranking input.
[Details]
Added functionality for FLITSR to be able to read in both Gzoltar
rankings and its own FLITSR ranking, without the need for the user to
specify which ranking is given. It does so by opening the file and
determining which type of ranking has been given, before actually
reading it. Any of the output functions can then be run on these
rankings.

[File's changed]

	modified:   flitsr.py
	modified:   ranking.py