Cleaning up the FLITSR implementation.
[Details]
This commit removes a few old and unused methods, as well as starting to
move things around. In future commits the implementation will be
refactored to be neater and run better.

[Bugs]
Fixed a small bug preventing FLITSR from continuing to look in the
ranking if the top-most element is not executed in failing tests. This
will help metrics like barinel that do not adhere to the one rule of
rankings (that elements at the top fail more tests than elements below
them).

[File's changed]
	modified:   flitsr.py
	modified:   input.py
	modified:   output.py
	modified:   tcm_input.py