Began refactoring spectrum table.
[Details]
As a part of the flitsr refactoring, the table used to hold the spectrum
for the program needs to be refactored to be an object. This will help
with the program understanding and further usage. The object that will
be used for the spectrum is created in this commit, but still needs work
before it can be used.

[File's changed]

	new file:   spectrum.py