Fixed python package problems.
[Details]
The use case for flitsr is to be usable as a command line tool AND as a
python module (for the advanced users). In order to do this, the flitsr
package used to have some python path magic within the module, causing
some errors in its interactions with tools such as mypy. To fix these,
the proper python format was followed, and instead the PYTHONPATH is now
used to expose FLITSR. This is a happy compromise between having flitsr
available everywhere and conforming to python conventions.

[File's changed]
	modified:   bin/flitsr
	modified:   bin/merge
	modified:   bin/percent_at_n
	modified:   bin/plot
	new file:   bin/ranking_stats
	modified:   bin/reduce
	modified:   bin/run_all
	modified:   flitsr/__init__.py
	modified:   flitsr/__main__.py
	modified:   flitsr/cutoff_points.py
	modified:   flitsr/flitsr.py
	modified:   flitsr/get_metric_names.py
	modified:   flitsr/input.py
	modified:   flitsr/localize.py
	modified:   flitsr/merge.py
	modified:   flitsr/parallel.py
	modified:   flitsr/percent_at_n.py
	modified:   flitsr/precision_recall.py
	modified:   flitsr/reduce.py
	modified:   flitsr/split_faults.py
	modified:   flitsr/tcm_input.py
	modified:   setup.sh
	deleted:    tests/__init__.py