Removed matplotlib and pytest requirement.
[Details]
Removed the user requirements of matplotlib and pytest by moving the
usage of these packages to only related parts of the system. The pytest
tests are moved to the tests directory and the pytest requirement to
test_requirements.txt. The plot script which uses matplotlib is now not
imported, allowing flitsr to be used without matplotlib.

[Note]
This may cause issues if users want to use the plot script.

[File's changed]
	modified:   flitsr/__init__.py
	modified:   flitsr/suspicious.py
	modified:   requirements.txt
	new file:   test_requirements.txt
	new file:   tests/test_suspicious.py