Hotfix: Fixed the run_all script to use venv.
[Bugs]
The python virtual environment changes added not too long ago did not
include the one call in the run_all script to a flitsr related module.
This would have made the run_all script unusable with the new changes.
This is now fixed so that the run_all script should work.

[File's changed]
	modified:   bin/run_all