Updated RC configuration for setup script.
[Details]
Changed the setup script to always update the FLITSR setup in the RC
file, but to preserve the RC file structure when doing so. These
changes will check for FLITSR_HOME in the RC file, and if found, will
replace these lines with the flitsr setup lines. If not found, the
lines will be placed at the end of the file.

[File's changed]
	modified:   setup.sh