Updated lauc calculation to be more useful.
[Details]
Updated the lauc calculation to prioritize the elements at the top of
the ranking. This is done by a somewhat 'reversed' log function that
increases slowly at first and then rapidly. This is different to an
exponential function in that the rate is much slower at first. This lauc
calculation is much more useful for evaluation purposes.

[File's changed]
	modified:   flitsr.py