Updated merge script to be more generic.
[Details]
The merge script was currently hardcoded for the particular
configuration which was given by the run_all script and 'all' argument
to flitsr. This however restricts the techniques and calculations that
are used to only those that are detailed in these places. To make the
merge script more general, the script was completely rewritten. The
rewrite now dynamically supports any metrics, modes and calculations
that it is given (and also fixes a few issues in the old script). This
merge script should now be useful for any experimentation that produces
.results files.

[File's changed]
	modified:   __init__.py
	modified:   file.py
	modified:   merge.py