Updated the run_all script.
[Details]
Modified the run_all script to be more generic (like the merge script in
a previous commit). This includes stability changes so that the run_all
script can be run on any type of directory structure. The script now has
options to specify the depth, and/or the type (TCM or Gzoltar) of the
input files.

[Note]
The script has the option to specify the metrics to use, however this
feature has not been implemented yet.

[File's changed]
	modified:   bin/run_all
	modified:   flitsr/flitsr.py
	modified:   flitsr/merge.py