Added TCM ext and fixed bugs in run_all script.
[Details]
Added the TCM optional argument for the TCM file's extension. This
allows the script to look only for a particular file type when searching
for TCMs, avoiding other files that may be in the directories. Note that
this argument is optional, and not giving the extension will still be
valid.

[Bugs]
Fixed a bug related to having multiple directories for the run_all
script.

[File's changed]
	modified:   run_all