Continued refactoring using Spectrum, added score
[Details]
Continued the refactoring using the Spectrum object. This included
finishing editing the Gzoltar input method to use spectrum, as well as
changing the TCM input method to use spectrum as well. Also re-worked
the rest of the flitsr class and the output file. In order to facilitate
the new Spectrum object (and better readability), the new Score and
Scores objects were created. The Score object holds a scored element
according to an SBFL method, and the Scores object holds multiple Score
objects. Changed the implementation to use the Scores and Score object.
With the addition of the Scores object, the localize class could be
discarded, and its components spread between the Scores and suspicious
classes.

[File's changed]
	modified:   flitsr.py
	modified:   input.py
	deleted:    localize.py
	modified:   output.py
	modified:   parallel.py
	new file:   score.py
	modified:   spectrum.py
	modified:   suspicious.py
	modified:   tcm_input.py