Efficiency changes in Spectrum and fixed bugs.
[Details]
The Spectrum object underwent major efficiency improvements to increase
the objects space and time efficiency. Before this commit, the Spectrum
object would occupy large amounts of memory on big spectra, ultimately
crashing, and would also take too long. These improvements make the
object both faster and smaller.

To make it smaller, a bitarray object is
used which stores the spectrum as bits instead of bytes in a list. In
addition the elements object is removed from each execution, and
instead, each execution has a reference to the Spectrum's global
elements list.

To make the implementation faster, the global elements list was changed
to be a Dict, which supports fast lookup, but still has almost all the
same properties of the list (on the Dict's keys, which are the elements).
In addition, each element item now has the details object split up,
which provides a better interface, but also a faster hash function.

[Bugs]
The input methods were both fixed to be in-line with the current
Spectrum object.

[File's changed]
	modified:   flitsr/input.py
	modified:   flitsr/spectrum.py
	modified:   flitsr/suspicious.py
	modified:   flitsr/tcm_input.py
	modified:   requirements.txt