Moved Spectrum operations from input to Spectrum.
[Details]
Changed the use of the Spectrum object to be more API-like, such that
the operations performed on the Spectrum object are done within the
Spectrum and not outside. In particular, the incrementing of the counts
is now handled inside the Spectrum object's methods, instead of inside
the input classes.

[Note]
Also changed the Gzoltar input method to only add executed elements to
the Spectrum to save on time.

[File's changed]
	modified:   input.py
	modified:   spectrum.py
	modified:   tcm_input.py