Added tests for input methods.
[Details]
Updated the test suite to fix a few errors and to add more tests for the
input methods (Gzoltar and TCM). The input methods now have randomized
stress-tests to ensure the correctness of the input readers.

[Bugs]
Fixed a minor bug in the suspicious class due to the addition of the new
apply_formula method.

[File's changed]
	modified:   flitsr/suspicious.py
	modified:   test_requirements.txt
	new file:   tests/helper.py
	modified:   tests/input_test.py
	renamed:    tests/test_suspicious.py -> tests/suspicious_test.py
	new file:   tests/tcm_input_test.py