Added ouput file and csv options.
[Details]
Added the output ('-o') option to enable flitsr to print directly to a
file. The framework for this was already in place due to the run all
mode in flitsr.
Also added the csv output option to flitsr, which makes flitsr output
the ranking in Gzoltar csv format instead of the usual flitsr format.

[File's changed]
	modified:   args.py
	modified:   flitsr.py
	modified:   output.py
	modified:   spectrum.py