Refactored argument handling, allow many metrics.
[Details]
Refactored the argument handling, in particular moving all argument
handling that was originally in the flitsr.main function to the args
module. The refactoring also allowed the all and default modes to be
done in a single format in the main function.

Added functionality for multiple metrics to be specified for flitsr.
When multiple metrics are specified in the normal mode, the outputs are
stored in separate files. When multiple metrics are specified in all
mode, these override the default metrics used for all.

The all mode now also supports overriding of output types.

Added all implemented metrics to the list of available metrics to run.
In addition, a curated set of metrics is also available which is used by
default for certain functions.

[Note]
- Added the Flitsr_Types Enum class that stores the various supported
  flitsr mode types (base sbfl, flitsr, and flitsr multi)
- The argument handling now checks for correct input files

[File's changed]
	modified:   flitsr/args.py
	modified:   flitsr/flitsr.py
	new file:   flitsr/flitsr_types.py
	modified:   flitsr/suspicious.py