Improved the efficiency of the Scores object.
[Details]
The scores object was initially taking unusually long to sort using the
tiebrk=3 strategy. This was determined to be because of the __getitem__
method in the scores object, which was used to get the original scores
of elements to sort, and whose efficiency was linear in the amount of
elements. To remedy this, the __getitem__ function was re-written using
a new dictionary object that stores scores for each element. This
dictionary now has O(1) timing, but needs to be maintained in
conjunction with the scores list. To do this, the scores object was
made private, and all references to it outside of the Scores class were
replaced with helper functions which also modify the dictionary.

[File's changed]
	modified:   cutoff_points.py
	modified:   score.py