Added run_all recovery mechanism.
[Details]
Added a recovery mechanism for the run_all function via the -r option.
Specifying this option to the run_all function will allow it to reuse
existing files that were potentially made available in a previous
(failed) run. Also added the --no-override option to flitsr which ensures
files are skipped and not overridden by flitsr, which facilitates part of
the recovery mechanism.

[File's changed]
	modified:   ../bin/run_all
	modified:   args.py
	modified:   flitsr.py