Spectrum retains execution of removed elements.
[Details]
Added the ability of the spectrum to retain the execution information of
removed elements. This is particularly useful for elements that are
removed as part of groupings, which are still necessary as part of the
spectrum but have identical spectra to the rest of the group. In order
to do the retention, the spectrum now has a list of current spectral
elements as well as a dictionary of all the elements to have ever been
in the spectrum.

[File's changed]
	modified:   flitsr/spectrum.py