Made merge script print predictable ordering.
[Details]
The previous version of the merge script had functionality allowing it
to be general in the metrics, modes, and calculations that it merges.
This however made the results be printed out in a random order, which
made it hard to inspect results files, particularly comparing different
ones. This commit fixes this issue by printing these inputs in a
predictable order. The order is determined by sorting the lists of
inputs.

[File's changed]
	modified:   merge.py