Improved the style of the perc@n plots.
[Details]
Changed the layout of the legend for the percent-at-n plots. The legend
now has separate entries for the type of metric, and type of "mode"
(i.e. base, flitsr, multi, etc.). This makes it much easier to see the
colour belonging to each metric, and the line style for each mode.

Also added more line styles to accommodate other mode types (such as
parallel or Artemis if needed).

[File's changed]
	modified:   percent_at_n.py