Added error storing to run_all script.
[Details]
Due to the recent addition of the progress bar in the run_all script,
errors printed to the terminal from the FLITSR command do not work as
well. This is increased by the warning messages which FLITSR may
regularly produce. To remedy this, these error messages are now stored
in a results.err file on a per project and per file basis. This also
allows easier reading of the error messages after the run.

[File's changed]
	modified:   bin/run_all