Updated the Ties class and corresp. calculations.
[Details]
Updated the Ties class to:
1. Not use worst_effort anymore
2. Accurately calculate the number of fault locations in a group
3. Manage the collapse argument through various helper functions that
   return either information pertaining to element level or group level.

Also updated the various calculations to respect this new ties object,
which removes the need for explicitly handing the collapse cases.

[Bugs]
There may be a variety of bugs in this commit considering this is a
large change to the calculations. This will be investigated and
potentially fixed in following commits.

One known bug in the system is with the wasted effort calculations,
which currently do not fully support multi-location multiple faults.

[File's changed]
	modified:   flitsr/flitsr.py
	modified:   flitsr/percent_at_n.py
	modified:   flitsr/precision_recall.py
	modified:   flitsr/score.py
	modified:   flitsr/top.py
	modified:   flitsr/weffort.py