Added partial saving of run_all error messages.
[Details]
Previously the run_all script only saved the error messages from each
flitsr run to the results.err file at the end of the script, which meant
that if the script exited early, no error messages were saved. This is a
problem for PC crashes and when manually killing the script. To remedy
this, functionality was added to save parts of the error messages for
each project/directory seperately, before the end of the script.

[Bugs]
- Fixed a bug in FLITSR for when it is given an incorrectly formatted
  file. Previously FLITSR would silently ignore this due to a bug in
  checking for this case, and would compute stats for this file anyway.
  This commit fixes this and ensures an error message is printed out.

[File's changed]
	modified:   bin/run_all
	modified:   flitsr/flitsr.py