Updated spectrum iteration and failing tests.
[Details]
Updated the behaviour of the spectrum and individual execution
iteration. When the spectrum is iterated over, it now returns each
execution, and when an execution is iterated over, it returns the
execution values for each element (that is currently in the spectrum).
This behaviour now reflects the spectrum being a matrix of executions.

Also made the spectrum's failing tests a private field with a getter
method, which enhances encapsulation.

[File's changed]
	modified:   flitsr.py
	modified:   output.py
	modified:   spectrum.py
	modified:   split_faults.py