Moved the tests_executing method to the spectrum.
[Details]
Moved the tests_executing method that was in the flitsr.py module to the
spectrum and called it get_tests. This method is currently used to get
the failing tests executing a given element in the spectrum, and is used
mostly in the FLITSR algorithm. The new get_tests method in the spectrum
can now be used to get all (or just the failing) tests executing a given
element, and is more logical to be in the spectrum class.

[Note]
Also made a slight change to the line separator for multiple spectrums
in FLITSR's output. This should have no logical difference though.

[File's changed]
	modified:   flitsr.py
	modified:   spectrum.py