Minor stability improvements to spectrum matrix.
[Details]
Added minor improvements to the to_matrix method of the Spectrum class,
which should improve the stability of the method. Also added custom
equality checks for the test class, which should also improve stability
of the flitsr framework.

[Note]
The ARTEMIS algorithm uses randomness which causes the results to be
different for each run. This makes comparison of the results difficult
and somewhat impossible.

[File's changed]
	modified:   spectrum.py