Fixed bug in FLITSR internal ranking.
[Details]
Fixed a bug in the FLITSR internal ranking (i.e. ranking of basis
elements), and in doing so, discovered that ranking these elements in
reverse order to what is in the ISSTA paper, so that now the last
element found by FLITSR is first, is better. Description of the bug
found is below.

[Bugs]
- The bug assigned ranks to the basis elements based on their position
  in the original ranking and not the FLITSR basis, which made the ranks
  of basis elements somewhat random.

[File's changed]
	modified:   flitsr.py