Added FLITSR internal ranking ordering option.
[Details]
Added an option to specify the method to use when ranking the elements
of a FLITSR basis. By default FLITSR will use what was active
previously, which is the "flitsr" option. This option gives the basis in
the order FLITSR returns it (i.e. on the way out of its recursion
chain).

[File's changed]
	modified:   README.md
	modified:   flitsr/args.py
	modified:   flitsr/flitsr.py