Complete overhaul of the merge script.
[Details]
This commit includes a complete overhaul of the merge script's API and
internal workings. The overhaul includes the following:
1. A new API interface using the python argparse module
2. A redesign of the merge script to facilitate flexibility
3. Many new, useful commands implemented in the redesigned script.

The updated API is also now available in the README.

[Note]
The general usage of the merge script has not changed much, i.e.
executing the command `merge` on its own will still produce roughly the
same result. The differences in the API affect mainly the advanced options
for merge.

[File's changed]
	modified:   README.md
	modified:   bin/run_all
	modified:   flitsr/merge.py