Actual fix for precision/recall calculations.
[Details]
The previous bug fixes for the precision/recall values attempted to
'patch' the errors by just ensuring the program does not crash. This
commit updates the calculations to now actually report accurate numbers.
The calculation now respects the number of locations and number of
faults being different, and factors this in the the result.

[File's changed]
	modified:   precision_recall.py