Added the --types argument for enhanced types.
[Details]
Added the -t/--types aregument to flitsr to give enhanced control over
the FLITSR advanced types that are used. The option overrides all other
individual advanced types options, but these options are still available
if the --types option is not used. The --types option can also be
specified multiple times to run multiple advanced type combinations.
This is particularly useful for the --all mode.

[File's changed]
	modified:   flitsr/args.py