Updated requirements and added completion.
[Details]
Updated the FLITSR requirements file to contain requirements necessary
for python 3.12 (and also to be backwards compatible).

Also added command completion for the flitsr commands through the use of
the python argcomplete module. This is automatically installed and set
up using the setup.sh script as well. The completion is only enabled for
the main flitsr command for now.

[File's changed]
	modified:   flitsr/args.py
	modified:   requirements.txt
	modified:   setup.sh