Fixed issue with merge incl/excl functionality.
[Details]
Fixed the issue described here: https://trello.com/c/eHyqhFqw, which
relates to the new include and exclude options for the merge script.
These options now optionally take a depth in addition to each file name,
which specifies at which depth that option is valid for. A depth of '*'
may also be given, and is the default, indicating any depth is valid.
The include and exclude options are now made to consider directories
other than at the last level of recursion.

[File's changed]
	modified:   merge.py