Added TCM output and parallel method & Gzoltar.
[Details]
Added a new output option which allows printing of a given spectrum
into TCM format. This is useful if the read in TCM was originally in
Gzoltar format, or if the spectrum was altered.

The above is then used in the new handling of Gzoltar and method level
spectra for the parallel technique. FLITSR now supports the parallel
technique for any input given, and will do so by writing out a TCM file
if there is none already.

[File's changed]
	modified:   flitsr/flitsr.py
	modified:   flitsr/output.py
	modified:   flitsr/parallel-1.0-SNAPSHOT-jar-with-dependencies.jar
	modified:   flitsr/parallel.py