Added input error catching & error class.
[Details]
Added a new class called errors.py which deals with printing out errors
and warnings, and used this in the input reading.

Added more advanced input error catching and reporting to prevent silent
input bugs.

[Note]
The construction of the spectrum may require a full refactoring to
improve the stability of the code. This may come in a future commit.

[File's changed]
	new file:   flitsr/errors.py
	modified:   flitsr/input.py
	modified:   flitsr/output.py
	modified:   flitsr/spectrum.py
	modified:   flitsr/tcm_input.py