Overhaul of spectrum grouping and creation.
[Details]
Following the changes to the input error catching, this commit does a
complete overhaul of the spectrum creation and grouping components. The
spectrum is now no longer created in partial states, but is created
using a SpectrumBuilder, which constructs the spectrum all-at-once.
Because of this new design, it is now possible for the spectrum to have
a separation between groups and elements (collectively called Entities)
where groups are collections of elements and are the actual entries in
the spectrum. Almost all operations now work with groups in the
spectrum, except when the actual underlying elements are required.

[Note]
One known issue is with the ranking input, which doesn't produce a
spectrum in the conventional way. This needs to be fixed in a future
commit.

[Bugs]
There may be numerous bugs in this commit, but these will be ironed out
in a later commit.

[File's changed]
	modified:   artemis_wrapper.py
	modified:   cutoff_points.py
	modified:   flitsr.py
	modified:   input.py
	modified:   output.py
	modified:   parallel.py
	modified:   percent_at_n.py
	modified:   precision_recall.py
	modified:   ranking.py
	modified:   score.py
	modified:   spectrum.py
	new file:   spectrumBuilder.py
	modified:   suspicious.py
	modified:   tcm_input.py
	modified:   tie.py