Renamed Score(s) object to Rank/Ranking.
[Details]
Complete refactor of the Score and Scores objects renaming them to Rank
and Ranking respectively. This renaming makes the code much more
readable and intuitive, improving the usability of FLITSR.

Also moved the tie breaking option from the Ranking's sort() method to
the Ranking constructor. This allows subsequent calls to sort() without
needing to know the tie breaking method used for the ranking. The
constructor also has a default value for the tie breaking in cases where
this is unknown or irrelevant.

[File's changed]
	modified:   artemis_wrapper.py
	modified:   cutoff_points.py
	modified:   flitsr.py
	modified:   output.py
	modified:   ranking.py
	new file:   read_ranking.py
	deleted:    score.py
	modified:   suspicious.py
	modified:   tie.py