Started adding test selection for mutation.
[Details]
Added a non-working version of the test selection for mutation. This
process allows only the tests executing a particular element to be run
when doing the mutation testing check, which should significantly
improve performance.

[Note]
This is a non-working version, so it WILL crash when run.

[Bugs]
Currently the program crashes because the <<TEST>> identifier is used in
the constructor when running all tests and is not replaced.

[File's changed]
	modified:   flitsr.py
	modified:   mutate_wrapper.py