Minor bugFix for spectrumBuilder.
[Bugs]
The spectrumBuilder assumed that every test case executed at least one
element, and would produce an error related to the uninitialized
execution object if this was not the case. To rectify this, the
spectrumBuilder now initializes every tests execution when the test is
added, so that even if the test has no executions, no error is produced.

[File's changed]
	modified:   spectrumBuilder.py