Added drivers for running ordering experiments.
[Details]
Added the ordering_driver.py, which is a copy of the main method from
flitsr with slight changes, allowing it to run all of the various types
of basis ordering methods. Also created a duplicate of the run_all
script called run_ordering, which likewise has minor modifications
allowing it to run the experiments for the basis ordering methods.

[File's changed]
	new file:   ../bin/run_ordering
	new file:   ordering_driver.py