Made FLITSR installable as a python package.
[Details]
In this commit, all the rest of the changes were made to make FLITSR
properly installable as a python package. FLITSR can now be install
either with pip, or pipx, and comes with the command-line scripts and
python module import (only for pip) as it did before.

[Note]
The FLITSR package is currently on the test PyPi server, and works
installing from there. This means FLITSR is ready to be shipped as a
python package when given the go-ahead.

[File's changed]
	new file:   MANIFEST.in
	modified:   README.md
	modified:   bin/run_all
	modified:   flitsr/__main__.py
	modified:   flitsr/args.py
	modified:   flitsr/flitsr.py
	modified:   flitsr/merge.py
	modified:   flitsr/parallel.py
	modified:   flitsr/percent_at_n.py
	modified:   pyproject.toml