Added the option to run_all with another driver.
[Details]
Added the option for the run_all script to specify the driver to use
when running flitsr. By default, the driver used for the run_all script
is the main flitsr entry-point, however this can now be changed to be
any other driver/entry-point.

This option may now be used to run the flitsr basis ordering
experiments, by specifying the driver to be the custom ordering_driver
created for these experiments. An other experiments can now be run in
the same way, making it easy to create new experiments for the flitsr
evaluation framework, and run them using the run_all script.

[File's changed]
	modified:   run_all
	deleted:    run_ordering